Finalizando a aula, implementando recursos http para buscar notes do mocky.io e usando pattern observable

+++ src/pages/detail/detail.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
-import { NoteService } from '../../app/models/note.service';
+import { IonicPage, NavController, NavParams, AlertController } from 'ionic-angular';
+import { NoteService } from '../../app/services/note.service';
+import { Note } from '../../app/models/note.model';
 
 /**
  * Generated class for the DetailPage page.
@@ -16,19 +17,51 @@ import { NoteService } from '../../app/models/note.service';
 })
 export class DetailPage {
 
-  note = [];
+  note: Note;
+  newNoteFlag: boolean = false;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, private noteService: NoteService) {
+  constructor(
+    public navCtrl: NavController,
+    public navParams: NavParams,
+    private noteService: NoteService,
+    private alertCtrl: AlertController) {
+
+    this.note = this.navParams.get("noteParam");
+    if (!this.note) {
+      this.note = new Note();
+      this.newNoteFlag = true;
+    }
+    
   }
 
   onTrash() {
-    this.noteService.removeNote(this.note);
-    this.navCtrl.pop();
+    //constrói o alerta
+    let confirm = this.alertCtrl.create({
+      title: "Delete?",
+      message: `Are you sure you want to delete this note: "${this.note.title}"?`,
+      buttons: [
+        //primeiro botão, sem handler não faz nada
+        {
+          text: "Cancel"
+        },
+        //segundo botão
+        {
+          text: "Confirm",
+          handler: () => {
+            this.noteService.removeNote(this.note);
+            this.navCtrl.pop();
+          }
+        }
+      ]
+    });
+    //exibe
+    confirm.present();
   }
 
-  ionViewDidLoad() {
-    this.note = this.navParams.get("noteParam");
-    console.log("nav - param", this.note);
-  }
+  ionViewDidLoad() {}
 
+  ionViewWillLeave() {
+    if (this.newNoteFlag)
+      this.noteService.addNote(this.note);
+  }
 }
